2.3.1 (Redirect) * IPN - Implement Magento 2.3.2 compatibility with better Magento 2.2.x backwards compatibility. * Use order grand total price when sending total to PayGate for improved multi-currency support.
