+++ Redirect/Paygate/Paygate/composer.json
@@ -16,7 +16,7 @@
       "magento/magento-composer-installer": "*"
     },
     "type": "magento2-module",
-    "version": "2.3.0",
+    "version": "2.3.1",
     "license": [
         "OSL-3.0",
         "AFL-3.0"
